Rename discord/params to discord/config
Just doing this for consistency with the other plugins.

Test plan: Simple rename-refactor, `yarn flow` passes.
